appFrame {appFrame}	R Documentation

Display NOAA header and footer for framing Shiny apps

Description

Display scrolling or non-scrolling NOAA banner header and footer elements for framing Shiny apps.

Usage

appFrameFooterFixed(displayAppsURL="https://dataexplorer.nwfsc.noaa.gov",
    contactURL="https://www.nwfsc.noaa.gov/contact/feedback.cfm")
appFrameFooterScrolling(displayAppsURL="https://dataexplorer.nwfsc.noaa.gov",
    contactURL="https://www.nwfsc.noaa.gov/contact/feedback.cfm")
appFrameHeaderFixed(overlapHeight=7, includeBanner=FALSE)
appFrameHeaderScrolling(overlapHeight=7, includeBanner=FALSE)

Arguments

contactURL
URL in the "Problems with site?" footer link.

displayAppsURL
URL in the "Display apps" footer link.

overlapHeight
height in em of blank div that lies beneath header to prevent normally-flowing content from being hidden by header. The larger the value of overlapHeight, the farther down the page the content will appear.

includeBanner
if TRUE, display contents of banner.html immediately below header logos; if FALSE, do not display banner.

Details

These functions return lists of HTML tags which render the NOAA banner header and contact links footer for display above and below Shiny app interfaces. appFrameHeaderFixed and appFrameFooterFixed render elements that do not scroll with the contents of the browser window, while headers and footers created with appFrameHeaderScrolling and appFrameFooterScrolling will scroll along with the rest of the browser window contents.

To display a banner message beneath the header, set includeBanner to TRUE. The contents of banner.html will be displayed beneath the header logos. Current banner contents announce the changeover of the site from HTTP to HTTPS. Changing the banner requires editing banner.html and rebuilding the package.

If a banner is displayed, overlapHeight will probably need to be increased.

The functions have so far been tested with navbarPage and fluidPage-embedded sidebarLayout layout schemes. Using sidebarLayout, appFrameHeaderFixed should be included as an argument to fluidPage before sidebarLayout and appFrameFooterFixed should be included as an argument following the end of the sidebarLayout definition. This results in a header that remains visible at the top of the browser window and a footer that remains visible at the bottom when the window is scrolled.

Non-scrolling headers don't work properly in navbarPage layouts; to display a scrolling header and a fixed footer, pass appFrameHeaderScrolling() as the header argument to navbarPage and appFrameFooterFixed() as the footer argument. This results in a header element displayed below the navbarPage menu but above the contents of each tab, which scrolls with the contents of the tab. The footer remains visible at the bottom of the browser window.

Value

Each function returns a tagList containing the HTML and CSS elements necessary for rendering its element.

Examples

## Not run: 
library(appFrame)
fluidPage(
  appFrameHeaderFixed()
  sidebarLayout(...)
  appFrameFooterFixed(displayAppsURL="http://apps.home.page/url")
)

library(appFrame)
navbarPage(
  "NavBar Title",
  tabPanel(...),
  ...,
  header=appFrameHeaderScrolling(),
  footer=appFrameFooterScrolling(contactURL="http://report.a.problem/issues")
)

## End(Not run)

[Package appFrame version 0.1.8 Index]